refactor: extract ExerciseQueue component - address Sandy Metz feedback
- Extract comprehensive exercise queue UI (planned sets, completed sets, navigation)
- Create ExerciseQueue component (206 lines) with clear props interface
- Reduce main component from 644 to 490 lines (-154 lines)
- Include: planned sets preview, Log All Sets, completed sets list, exercise navigation
- Total reduction: 867 → 490 lines (-377 lines, -43% reduction\!)
- Major milestone: Under 500 lines, addressing Sandy's massive component violation
- Part of issue #27 Sandy Metz review feedback
