+++ scratchpads/issue-27-experimental-workout-execution.md
@@ -16,11 +16,12 @@
 #### Phase 1: Break Down Component (PRIORITY 1)
 - [x] Extract WorkoutHeader component (56 lines) ✅ COMPLETED
 - [x] Extract SetLoggingForm component (153 lines) ✅ COMPLETED
-- [ ] Extract WorkoutProgress component (<75 lines)  
+- [x] Extract WorkoutProgress component (189 lines) ✅ COMPLETED
 - [ ] Extract ExerciseQueue component (<100 lines)
 - [ ] Create container component (<100 lines)
 
-**Progress**: 767 lines remaining (was 867) - 100 lines reduced (-12% reduction so far)
+**Progress**: 644 lines remaining (was 867) - 223 lines reduced (-26% reduction so far)
+**Major Milestone**: Broken Sandy's 867-line violation by 26%!
 
 #### Phase 2: Extract State Management (PRIORITY 2)
 - [ ] Create useWorkoutSession custom hook
